+++ src/main/java/site/licsber/shop/dto/IndexItemDTO.java
@@ -3,7 +3,7 @@ package site.licsber.shop.dto;
 import lombok.Data;
 import site.licsber.shop.model.entity.ItemCategory;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Data
 public class IndexItemDTO {
@@ -11,11 +11,11 @@ public class IndexItemDTO {
     private Integer id;
 
     private String title;
-    private BigInteger price;
-    private BigInteger originPrice;
+    private BigDecimal price;
+    private BigDecimal originPrice;
 
     // 邮费
-    private BigInteger postage;
+    private BigDecimal postage;
     private String primaryImg;
 
     // 0-实物商品 1-电子商品
